requirements-dev.txt
pip
pytest
codecov
requests-mock
doctr
sphinx
pre-commit
pip
flaky
pytest-xdist
pytest-cov
setuptools_scm>=7
python-build
